Match also vmlinux* in Debian live example
The Linux kernel from the linux-image-powerpc64le package on ppc64el is
called `/boot/vmlinux-${version}-powerpc64le`.

So relax the glob for copying the Linux kernel to match `/boot/vmlinux*`
in addition to `/boot/vmlinuz*`. Keep the name `vmlinuz` for the copy
target in both cases to keep the example simple.

Signed-off-by: Benjamin Drung <bdrung@posteo.de>